requirements.txt
certifi==2022.9.24
charset-normalizer==3.1.0
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
idna==3.4
influxdb-client==1.33.0
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
reactivex==4.0.4
requests==2.29.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12
